Handle systems with more than 2147483647 kB of memory (#704593).
Our code had a limit of 2TB for the memory check.  At least one reporter is
trying to install on a system with 3TB of memory and anaconda reports that
he does not have enough memory.  The log file reports he has -1119974932 kB
of memory.

Make two changes here.  The first is in iutil.py to explicitly use the
long() Python type for memory reporting.  The second is in the
totalMemory() function in loadermisc.c, which is the source of error
reported in #704593.

+++ iutil.py
@@ -438,7 +438,7 @@ def memInstalled():
             mem = fields[1]
             break
 
-    return int(mem)
+    return long(mem)
 
 ## Suggest the size of the swap partition that will be created.
 # @param quiet Should size information be logged?
